refactor: move remaining commands to own directories
