+++ src/check/check.ts
@@ -4,10 +4,10 @@ import {
   getGitHubToken,
   readSkillLock,
   type SkillLockEntry,
-} from './lib/lock/index.ts';
-import { checkContextUpdates, updateContext } from './lib/install/index.ts';
-import { track } from './lib/telemetry.ts';
-import { RESET, DIM, TEXT } from './lib/utils.ts';
+} from '../lib/lock/index.ts';
+import { checkContextUpdates, updateContext } from '../lib/install/index.ts';
+import { track } from '../lib/telemetry.ts';
+import { RESET, DIM, TEXT } from '../lib/utils.ts';
 
 // ============================================
 // Shared skill update detection
