+++ src/check/index.ts
@@ -1,0 +1,8 @@
+export {
+  runCheck,
+  runUpdate,
+  findSkillUpdates,
+  type SkillUpdate,
+  type SkillCheckError,
+  type SkillCheckResult,
+} from './check.ts';
