+++ src/cli.ts
@@ -4,14 +4,14 @@ import { readFileSync } from 'fs';
 import { join, dirname } from 'path';
 import { fileURLToPath } from 'url';
 import { runAdd, parseAddOptions, initTelemetry } from './add/index.ts';
-import { runCheck, runUpdate } from './check.ts';
+import { runCheck, runUpdate } from './check/index.ts';
 import { CommandError } from './lib/command-result.ts';
-import { runFind } from './find.ts';
-import { runInit } from './init.ts';
-import { runInstallFromLock } from './restore.ts';
-import { runList } from './list.ts';
-import { removeCommand, parseRemoveOptions } from './remove.ts';
-import { runSync, parseSyncOptions } from './sync.ts';
+import { runFind } from './find/index.ts';
+import { runInit } from './init/index.ts';
+import { runInstallFromLock } from './restore/index.ts';
+import { runList } from './list/index.ts';
+import { removeCommand, parseRemoveOptions } from './remove/index.ts';
+import { runSync, parseSyncOptions } from './sync/index.ts';
 import { RESET, BOLD, DIM, TEXT, YELLOW } from './lib/utils.ts';
 
 const __dirname = dirname(fileURLToPath(import.meta.url));
