+++ src/find/index.ts
@@ -1,0 +1,1 @@
+export { runFind, formatInstalls, searchSkillsAPI, type SearchSkill } from './find.ts';
