+++ src/init/index.ts
@@ -1,0 +1,1 @@
+export { runInit } from './init.ts';
