+++ src/remove/index.ts
@@ -1,0 +1,1 @@
+export { removeCommand, parseRemoveOptions, type RemoveOptions } from './remove.ts';
