+++ src/restore/index.ts
@@ -1,0 +1,1 @@
+export { runInstallFromLock } from './restore.ts';
