+++ src/restore/restore.ts
@@ -1,14 +1,14 @@
 import * as p from '@clack/prompts';
 import pc from 'picocolors';
-import { readLocalLock } from './lib/lock/index.ts';
-import { runAdd } from './add/add.ts';
-import { runSync, parseSyncOptions } from './sync.ts';
-import { getUniversalAgents } from './lib/agents/index.ts';
-import { readDotaiLock, getLockEntriesByType } from './lib/lock/index.ts';
-import { addPrompts, addAgents } from './lib/install/index.ts';
-import { parseSource } from './lib/parsers/index.ts';
-import { cloneRepo, cleanupTempDir } from './lib/git/index.ts';
-import type { LockEntry, TargetAgent } from './lib/types.ts';
+import { readLocalLock } from '../lib/lock/index.ts';
+import { runAdd } from '../add/add.ts';
+import { runSync, parseSyncOptions } from '../sync/sync.ts';
+import { getUniversalAgents } from '../lib/agents/index.ts';
+import { readDotaiLock, getLockEntriesByType } from '../lib/lock/index.ts';
+import { addPrompts, addAgents } from '../lib/install/index.ts';
+import { parseSource } from '../lib/parsers/index.ts';
+import { cloneRepo, cleanupTempDir } from '../lib/git/index.ts';
+import type { LockEntry, TargetAgent } from '../lib/types.ts';
 
 /**
  * Install all context from lock files:
