+++ src/sync/index.ts
@@ -1,0 +1,1 @@
+export { runSync, parseSyncOptions, type SyncOptions } from './sync.ts';
